+++ packages/adminui-vue/CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
  - UAutoField: in case attribute used in `attribute-name` prop is not exists in Repository a developer-friendly
    exception is throws instead of `can not read XXX of undefined`
+ - `USelectMultiple`: new prop `additionalButtons`, for add buttons to dropdown before button 'more'
 
 ### Changed